fix: move SearchFilterMatchCard and SearchFilterMatches to match/ subdirectory
